+++ packages/terra-demographics-banner/src/DemographicsBanner.jsx
@@ -94,26 +94,22 @@ const DemographicsBanner = ({
   intl,
   personName,
   ...customProps
-}) => {
-  const noDataProvided = intl.formatMessage({ id: 'Terra.demographicsBanner.noDataProvided' });
-
-  return (
-    <DemographicsBannerDisplay
-      {...customProps}
-      age={age || noDataProvided}
-      dateOfBirth={dateOfBirth || noDataProvided}
-      gender={gender || noDataProvided}
-      personName={personName || noDataProvided}
-      dateOfBirthLabel={intl.formatMessage({ id: 'Terra.demographicsBanner.dateOfBirth' })}
-      dateOfBirthFullText={intl.formatMessage({ id: 'Terra.demographicsBanner.dateOfBirthFullText' })}
-      deceasedDateLabel={intl.formatMessage({ id: 'Terra.demographicsBanner.deceased' })}
-      gestationalAgeLabel={intl.formatMessage({ id: 'Terra.demographicsBanner.gestationalAge' })}
-      gestationalAgeFullText={intl.formatMessage({ id: 'Terra.demographicsBanner.gestationalAgeFullText' })}
-      postMenstrualAgeLabel={intl.formatMessage({ id: 'Terra.demographicsBanner.postMenstrualAge' })}
-      postMenstrualAgeFullText={intl.formatMessage({ id: 'Terra.demographicsBanner.postMenstrualAgeFullText' })}
-    />
-  );
-};
+}) => (
+  <DemographicsBannerDisplay
+    {...customProps}
+    age={age}
+    dateOfBirth={dateOfBirth}
+    gender={gender}
+    personName={personName}
+    dateOfBirthLabel={intl.formatMessage({ id: 'Terra.demographicsBanner.dateOfBirth' })}
+    dateOfBirthFullText={intl.formatMessage({ id: 'Terra.demographicsBanner.dateOfBirthFullText' })}
+    deceasedDateLabel={intl.formatMessage({ id: 'Terra.demographicsBanner.deceased' })}
+    gestationalAgeLabel={intl.formatMessage({ id: 'Terra.demographicsBanner.gestationalAge' })}
+    gestationalAgeFullText={intl.formatMessage({ id: 'Terra.demographicsBanner.gestationalAgeFullText' })}
+    postMenstrualAgeLabel={intl.formatMessage({ id: 'Terra.demographicsBanner.postMenstrualAge' })}
+    postMenstrualAgeFullText={intl.formatMessage({ id: 'Terra.demographicsBanner.postMenstrualAgeFullText' })}
+  />
+);
 
 DemographicsBanner.propTypes = propTypes;
 DemographicsBanner.defaultProps = defaultProps;