+++ packages/terra-demographics-banner/src/_LargeDemographicsBannerDisplay.jsx
@@ -114,7 +114,7 @@ const LargeDemographicsBannerDisplay = (props) => {
         <div className={cx('info-tile-container')}>
           <div className={cx('info-tile')}>
             <span className={cx('info-text')}>{infoText}</span>
-            <span className={cx('info-icon')}/>
+            <span className={cx('info-icon')} />
           </div>
         </div>
       )}