+++ packages/terra-demographics-banner/src/terra-dev-site/doc/example/DemographicsBannerSmallContainer.jsx
@@ -12,7 +12,6 @@ const DemographicsBannerSmallContainer = () => (
     <DemographicsBanner
       age="25 Years"
       dateOfBirth="May 9, 1992"
-      deceasedDate="March 12, 2017"
       gender="True Male"
       identifiers={{ MRN: 12343, REA: '3JSDA' }}
       photo={<Image alt="My Cat" src={demographicsImage} />}