+++ packages/terra-demographics-banner/tests/jest/DemographicsBanner.test.jsx
@@ -25,6 +25,7 @@ it('renders the banner wrapper with all props', () => {
       gender="Male"
       gestationalAge="April 5, 2016"
       identifiers={{ MRN: 12343, REA: '3JSDA' }}
+      isConfidential
       photo=""
       personName="Johnathon Doe"
       postMenstrualAge="April 7, 2016"